Add files via upload
It is a CNN with two Dense layers: the first one with 32 neurons and second Dense layer with a Softmax activation function. The output is a one-hot vector indicating the possibility of distribution in 15 classes below: 

0, 'residential_area': 1, 'library': 2, 'home': 3, 'beach': 4, 'forest_path': 5, 'metro_station': 6, 'car': 7, 'bus': 8, 'tram': 9, 'cafe/restaurant': 10, 'park': 11, 'city_center': 12, 'grocery_store': 13, 'train': 14

Top_5_categorical_accuracy metric is used to define the accuracy and also predict the label. Final accuracy is around 36%. Still needs improving.